Chomp failure option value
This is explicitly to support a use case where the failure setting might
be externalized from the hiera.yaml file such that it can be turned on
and off by modifying an external file. E.g.

    options:
      failure: "%{::consul_failure_mode}"

And in site.pp:

    $consul_failure_mode = file('/etc/puppetlabs/puppet/consul_failure_mode.conf')

Adding the chomp here doesn't cost anything, and it makes the above
easier to implement.